+++ core/src/main/java/com/bumble/appyx/core/integration/NodeHost.kt
@@ -2,11 +2,10 @@ package com.bumble.appyx.core.integration
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.Stable
 import androidx.compose.runtime.State
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.runtime.saveable.mapSaver
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.platform.LocalLifecycleOwner
@@ -18,11 +17,7 @@ import com.bumble.appyx.core.integrationpoint.IntegrationPoint
 import com.bumble.appyx.core.modality.BuildContext
 import com.bumble.appyx.core.node.Node
 import com.bumble.appyx.core.node.build
-
-@Stable
-fun interface NodeFactory<N : Node> {
-    fun create(buildContext: BuildContext): N
-}
+import com.bumble.appyx.core.state.SavedStateMap
 
 /**
  * Composable function to host [Node].
@@ -31,15 +26,11 @@ fun interface NodeFactory<N : Node> {
  */
 @Composable
 fun <N : Node> NodeHost(
-    customisations: NodeCustomisationDirectory = NodeCustomisationDirectoryImpl(),
+    customisations: NodeCustomisationDirectory = remember { NodeCustomisationDirectoryImpl() },
     integrationPoint: IntegrationPoint,
     factory: NodeFactory<N>
 ) {
-    val node by rememberNode(factory, customisations)
-    DisposableEffect(integrationPoint) {
-        integrationPoint.attach(node)
-        onDispose { integrationPoint.detach() }
-    }
+    val node by rememberNode(factory, customisations, integrationPoint)
     DisposableEffect(node) {
         onDispose { node.updateLifecycleState(Lifecycle.State.DESTROYED) }
     }
@@ -58,28 +49,28 @@ fun <N : Node> NodeHost(
 @Composable
 internal fun <N : Node> rememberNode(
     factory: NodeFactory<N>,
-    customisations: NodeCustomisationDirectory
-): State<N> =
-    rememberSaveable(
+    customisations: NodeCustomisationDirectory,
+    integrationPoint: IntegrationPoint,
+): State<N> {
+
+    fun createNode(savedStateMap: SavedStateMap?): N =
+        factory
+            .create(
+                buildContext = BuildContext.root(
+                    savedStateMap = savedStateMap,
+                    customisations = customisations
+                ),
+            )
+            .build()
+            .apply { this.integrationPoint = integrationPoint }
+
+    return rememberSaveable(
         inputs = arrayOf(),
         stateSaver = mapSaver(
             save = { node -> node.saveInstanceState(this) },
-            restore = { state ->
-                factory.create(
-                    buildContext = BuildContext.root(
-                        savedStateMap = state,
-                        customisations = customisations
-                    ),
-                ).build()
-            },
+            restore = { state -> createNode(savedStateMap = state) },
         ),
     ) {
-        mutableStateOf(
-            factory.create(
-                buildContext = BuildContext.root(
-                    savedStateMap = null,
-                    customisations = customisations
-                )
-            ).build()
-        )
+        mutableStateOf(createNode(savedStateMap = null))
     }
+}