+++ sandbox/src/main/java/com/bumble/appyx/sandbox/client/interop/parent/RibsParentRib.kt
@@ -6,12 +6,12 @@ import com.badoo.ribs.core.customisation.RibCustomisation
 import com.badoo.ribs.core.modality.BuildParams
 import com.badoo.ribs.core.plugin.Plugin
 import com.badoo.ribs.core.view.ViewFactory
-import com.bumble.appyx.sandbox.client.interop.parent.RibsParentViewImpl.Factory
+import com.bumble.appyx.sandbox.client.interop.parent.RibsParentView.Factory
 
 interface RibsParentRib : Rib {
 
     class Customisation(
-        val viewFactory: Factory = Factory(),
+        val viewFactory: Factory = RibsParentViewImpl.Factory(),
         val name: String = "defaultName"
     ) : RibCustomisation
 }