+++ testing-unit-common/src/main/java/com/bumble/appyx/testing/unit/common/helper/InteractorTestHelper.kt
@@ -55,6 +55,7 @@ class InteractorTestHelper<N : Node>(
     ) : Node(buildContext = buildContext) {
         @Composable
         override fun View(modifier: Modifier) {
+            // Deliberately empty
         }
     }
 }