+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## Pending changes
 
 - [#39](https://github.com/bumble-tech/appyx/pull/39) – Added: Workflows implementation to support deeplinks
+- [#32](https://github.com/bumble-tech/appyx/pull/32) – Added: `BackPressHandler` plugin that allows to control back press behaviour via `androidx.activity.OnBackPressedCallback`
+- [#59](https://github.com/bumble-tech/appyx/issues/59) – Added: interface for ParentNodeView<>
+- [#32](https://github.com/bumble-tech/appyx/issues/69) – Added: Jetpack Compose Navigation code sample
+- [#81](https://github.com/bumble-tech/appyx/issues/81) – Added: Support integration point for multiple roots
+- [#65](https://github.com/bumble-tech/appyx/pull/65) – Added: `InteropBuilderStub` and `InteropSimpleBuilderStub` testing util classes
+- [#47](https://github.com/bumble-tech/appyx/issues/47) – Updated: The 'customisations' module is now pure Java/Kotlin.
+- [#85](https://github.com/bumble-tech/appyx/issues/85) – Updated: Improved `InteropView` error messaging when `Activity` does not implement `IntegrationPointAppyxProvider`
 
 
 ## 1.0-alpha03